+++ qmk_firmware/keyboards/keyball/keyball39/keymaps/kamidai/features/auto_click_layer.h
@@ -1,0 +1,145 @@
+////////////////////////////////////
+///
+/// 自動マウスレイヤーの実装 ここから
+/// 参考にさせていただいたページ
+/// https://zenn.dev/takashicompany/articles/69b87160cda4b9
+///
+////////////////////////////////////
+
+enum ball_state
+{
+  NONE = 0,
+  WAITING,   // マウスレイヤーが有効になるのを待つ。 Wait for mouse layer to activate.
+  CLICKABLE, // マウスレイヤー有効になりクリック入力が取れる。 Mouse layer is enabled to take click input.
+  CLICKING,  // クリック中。 Clicking.
+  SWIPE,     // スワイプモードが有効になりスワイプ入力が取れる。 Swipe mode is enabled to take swipe input.
+  SWIPING    // スワイプ中。 swiping.
+};
+
+enum ball_state state; // 現在のクリック入力受付の状態 Current click input reception status
+uint16_t click_timer;  // タイマー。状態に応じて時間で判定する。 Timer. Time to determine the state of the system.
+
+uint16_t to_reset_time = 800; // この秒数(千分の一秒)、CLICKABLE状態ならクリックレイヤーが無効になる。 For this number of seconds (milliseconds), the click layer is disabled if in CLICKABLE state.
+
+const int16_t to_clickable_movement = 0; // クリックレイヤーが有効になるしきい値
+const uint16_t click_layer = 6;          // マウス入力が可能になった際に有効になるレイヤー。Layers enabled when mouse input is enabled
+
+int16_t mouse_record_threshold = 30; // ポインターの動きを一時的に記録するフレーム数。 Number of frames in which the pointer movement is temporarily recorded.
+int16_t mouse_move_count_ratio = 5;  // ポインターの動きを再生する際の移動フレームの係数。 The coefficient of the moving frame when replaying the pointer movement.
+
+const uint16_t ignore_disable_mouse_layer_keys[] = {KC_LANG1, KC_LANG2}; // この配列で指定されたキーはマウスレイヤー中に押下してもマウスレイヤーを解除しない
+
+int16_t mouse_movement;
+
+// クリック用のレイヤーを有効にする。　Enable layers for clicks
+void enable_click_layer(void)
+{
+  layer_on(click_layer);
+  click_timer = timer_read();
+  state = CLICKABLE;
+}
+
+// クリック用のレイヤーを無効にする。 Disable layers for clicks.
+void disable_click_layer(void)
+{
+  state = NONE;
+  layer_off(click_layer);
+}
+
+//
+report_mouse_t pointing_device_task_user(report_mouse_t mouse_report)
+{
+  int16_t current_x = mouse_report.x;
+  int16_t current_y = mouse_report.y;
+
+  if (current_x != 0 || current_y != 0)
+  {
+
+    switch (state)
+    {
+    case CLICKABLE:
+      click_timer = timer_read();
+      break;
+
+    case CLICKING:
+      break;
+
+    case WAITING:
+      mouse_movement += my_abs(current_x) + my_abs(current_y);
+
+      if (mouse_movement >= to_clickable_movement)
+      {
+        mouse_movement = 0;
+        enable_click_layer();
+      }
+      break;
+
+    case SWIPE:
+      click_timer = timer_read();
+
+      if (my_abs(current_x) >= SWIPE_THRESHOLD || my_abs(current_y) >= SWIPE_THRESHOLD)
+      {
+        rgblight_sethsv(HSV_BLUE);
+        process_swipe_gesture(current_x, current_y);
+        is_swiped = true;
+
+        if (is_repeat == false)
+        {
+          state = SWIPING;
+        }
+      }
+      break;
+      ;
+
+    case SWIPING:
+      break;
+
+    default:
+      click_timer = timer_read();
+      state = WAITING;
+      mouse_movement = 0;
+    }
+  }
+  else
+  {
+    switch (state)
+    {
+    case CLICKING:
+      break;
+
+    case CLICKABLE:
+      if (timer_elapsed(click_timer) > to_reset_time)
+      {
+        disable_click_layer();
+      }
+      break;
+
+    case WAITING:
+      if (timer_elapsed(click_timer) > 50)
+      {
+        mouse_movement = 0;
+        state = NONE;
+      }
+      break;
+
+    case SWIPE:
+      rgblight_sethsv(HSV_RED);
+      break;
+
+    case SWIPING:
+      if (timer_elapsed(click_timer) > 300)
+      {
+        state = SWIPE;
+      }
+      break;
+
+    default:
+      mouse_movement = 0;
+      state = NONE;
+    }
+  }
+  mouse_report.x = current_x;
+  mouse_report.y = current_y;
+
+  return mouse_report;
+}
