+++ qmk_firmware/keyboards/keyball/keyball39/keymaps/kamidai/keymap.c
@@ -16,9 +16,6 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include QMK_KEYBOARD_H
-#include "quantum.h"
-
 // コード表
 // 【KBC_RST: 0x5DA5】Keyball 設定のリセット
 // 【KBC_SAVE: 0x5DA6】現在の Keyball 設定を EEPROM に保存します
@@ -33,419 +30,13 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // https://github.com/kamiichi99/keyball/tree/main/qmk_firmware/keyboards/keyball/keyball39/keymaps/kamidai
 // https://shirogane-lab.com/collections/all
 
-enum custom_keycodes
-{
-  KC_MY_BTN1 = KEYBALL_SAFE_RANGE, // Remap上では 0x5DAF
-  KC_MY_BTN2,                      // Remap上では 0x5DB0
-  KC_MY_BTN3,                      // Remap上では 0x5DB1
-};
+#include QMK_KEYBOARD_H
+#include "quantum.h"
 
-// 自前の絶対数を返す関数。 Functions that return absolute numbers.
-int16_t my_abs(int16_t num)
-{
-  if (num < 0)
-  {
-    num = -num;
-  }
-  return num;
-}
-
-// 自前の符号を返す関数。 Function to return the sign.
-int16_t mmouse_move_y_sign(int16_t num)
-{
-  if (num < 0)
-  {
-    return -1;
-  }
-  return 1;
-}
-
-////////////////////////////////////
-///
-/// 自動マウスレイヤーの実装 ここから
-/// 参考にさせていただいたページ
-/// https://zenn.dev/takashicompany/articles/69b87160cda4b9
-///
-////////////////////////////////////
-
-enum ball_state
-{
-  NONE = 0,
-  WAITING,   // マウスレイヤーが有効になるのを待つ。 Wait for mouse layer to activate.
-  CLICKABLE, // マウスレイヤー有効になりクリック入力が取れる。 Mouse layer is enabled to take click input.
-  CLICKING,  // クリック中。 Clicking.
-  SWIPE,     // スワイプモードが有効になりスワイプ入力が取れる。 Swipe mode is enabled to take swipe input.
-  SWIPING    // スワイプ中。 swiping.
-};
-
-enum ball_state state; // 現在のクリック入力受付の状態 Current click input reception status
-uint16_t click_timer;  // タイマー。状態に応じて時間で判定する。 Timer. Time to determine the state of the system.
-
-uint16_t to_reset_time = 800; // この秒数(千分の一秒)、CLICKABLE状態ならクリックレイヤーが無効になる。 For this number of seconds (milliseconds), the click layer is disabled if in CLICKABLE state.
-
-const int16_t to_clickable_movement = 0; // クリックレイヤーが有効になるしきい値
-const uint16_t click_layer = 6;          // マウス入力が可能になった際に有効になるレイヤー。Layers enabled when mouse input is enabled
-
-int16_t mouse_record_threshold = 30; // ポインターの動きを一時的に記録するフレーム数。 Number of frames in which the pointer movement is temporarily recorded.
-int16_t mouse_move_count_ratio = 5;  // ポインターの動きを再生する際の移動フレームの係数。 The coefficient of the moving frame when replaying the pointer movement.
-
-const uint16_t ignore_disable_mouse_layer_keys[] = {KC_LANG1, KC_LANG2}; // この配列で指定されたキーはマウスレイヤー中に押下してもマウスレイヤーを解除しない
-
-int16_t mouse_movement;
-
-// クリック用のレイヤーを有効にする。　Enable layers for clicks
-void enable_click_layer(void)
-{
-  layer_on(click_layer);
-  click_timer = timer_read();
-  state = CLICKABLE;
-}
-
-// クリック用のレイヤーを無効にする。 Disable layers for clicks.
-void disable_click_layer(void)
-{
-  state = NONE;
-  layer_off(click_layer);
-}
-
-////////////////////////////////////
-///
-/// 自動マウスレイヤーの実装 ここまで
-///
-////////////////////////////////////
-
-////////////////////////////////////
-///
-/// スワイプジェスチャーの実装 ここから
-/// 参考にさせていただいたページ
-/// https://www.reddit.com/r/ploopy/comments/pbmrh3/qmk_options_on_ploopy_mouse_gestures/
-/// https://github.com/RobertCurry0216/qmk_firmware/blob/59f83aac8fa06010c3a1a0a53fcd453d96ce2f80/keyboards/ploopyco/trackball/keymaps/robertcurry0216/keymap.c
-///
-////////////////////////////////////
-
-// 変数
-int16_t current_keycode;
-const int16_t SWIPE_THRESHOLD = 10;
-bool is_swiped = false;
-bool is_repeat = false;
-
-// スワイプジェスチャーで何が起こるかを実際に処理する関数です
-// 上、下、左、右、スワイプなしの5つのオプションがあります
-void process_swipe_gesture(int16_t x, int16_t y)
-{
-  if (current_keycode == KC_LCMD)
-  {
-    if (my_abs(x) > my_abs(y))
-    { // 英かな切り替え
-      unregister_code(KC_LCMD);
-
-      if (x > 0)
-      { // swipe right
-        tap_code(KC_LANG1);
-      }
-      else
-      { // swipe left
-        tap_code(KC_LANG2);
-      }
-    }
-
-    if (my_abs(x) < my_abs(y))
-    { // 拡大と縮小
-      if (y > 0)
-      { // swipe down
-        tap_code(KC_MINUS);
-      }
-      else
-      { // swipe up
-        tap_code(KC_EQUAL);
-      }
-    }
-  }
-
-  if (current_keycode == KC_D)
-  { // カーソル移動
-    if (my_abs(x) > my_abs(y))
-    {
-      if (x > 0)
-      { // swipe right
-        tap_code(KC_RIGHT);
-      }
-      else
-      { // swipe left
-        tap_code(KC_LEFT);
-      }
-    }
-
-    if (my_abs(x) < my_abs(y))
-    {
-      if (y > 0)
-      { // swipe down
-        tap_code(KC_DOWN);
-      }
-      else
-      { // swipe up
-        tap_code(KC_UP);
-      }
-    }
-  }
-
-  if (current_keycode == KC_T)
-  { // フリック風
-    tap_code(KC_BSPC);
-
-    if (my_abs(x) > my_abs(y))
-    {
-      if (x > 0)
-      { // swipe right
-        tap_code(KC_S);
-      }
-
-      else
-      { // swipe left
-        tap_code(KC_W);
-      }
-    }
-
-    if (my_abs(x) < my_abs(y))
-    {
-      if (y > 0)
-      { // swipe down
-        tap_code(KC_V);
-      }
-      else
-      { // swipe up
-        tap_code(KC_H);
-      }
-    }
-  }
-}
-
-////////////////////////////////////
-///
-/// スワイプジェスチャーの実装 ここまで
-///
-////////////////////////////////////
-
-// マクロキーを設定
-bool process_record_user(uint16_t keycode, keyrecord_t *record)
-{
-  current_keycode = keycode;
-
-  switch (keycode)
-  {
-  case KC_MY_BTN1:
-  case KC_MY_BTN2:
-  case KC_MY_BTN3:
-  {
-    report_mouse_t currentReport = pointing_device_get_report();
-
-    // どこのビットを対象にするか。 Which bits are to be targeted?
-    uint8_t btn = 1 << (keycode - KC_MY_BTN1);
-
-    if (record->event.pressed)
-    {
-      // キーダウン時
-      // ビットORは演算子の左辺と右辺の同じ位置にあるビットを比較して、両方のビットのどちらかが「1」の場合に「1」にします。
-      // Bit OR compares bits in the same position on the left and right sides of the operator and sets them to "1" if either of both bits is "1".
-      currentReport.buttons |= btn;
-      state = CLICKING;
-    }
-    else
-    {
-      // キーアップ時
-      // ビットANDは演算子の左辺と右辺の同じ位置にあるビットを比較して、両方のビットが共に「1」の場合だけ「1」にします。
-      // Bit AND compares the bits in the same position on the left and right sides of the operator and sets them to "1" only if both bits are "1" together.
-      currentReport.buttons &= ~btn;
-      enable_click_layer();
-    }
-
-    pointing_device_set_report(currentReport);
-    pointing_device_send();
-    return false;
-  }
-
-  // 自動クリックレイヤーでLang1とLang2を押せるようにする
-  case KC_LANG1:
-  case KC_LANG2:
-  {
-    if (state == CLICKABLE)
-    {
-      if (record->event.pressed)
-      {
-        // キーダウン時
-        enable_click_layer();
-        return true;
-      }
-      else
-      {
-        // キーアップ時
-        disable_click_layer();
-      }
-    }
-  }
-
-  // クリックすると state が SWIPE になり、離したら NONE になる
-  // MOD系
-  case KC_LCMD:
-    if (record->event.pressed)
-    {
-      // キーダウン時
-      state = SWIPE;
-      register_code(keycode);
-    }
-    else
-    {
-      // キーアップ時
-      clear_mods();
-      disable_click_layer();
-    }
-    return false;
-
-  // TAP系（フリック風）
-  case KC_T:
-    if (record->event.pressed)
-    {
-      // キーダウン時
-      state = SWIPE;
-      tap_code(keycode);
-    }
-    else
-    {
-      // キーアップ時
-      disable_click_layer();
-    }
-    return false;
-
-  // TAP系（キーリピートあり）
-  case KC_D:
-    if (record->event.pressed)
-    {
-      // キーダウン時
-      state = SWIPE;
-      is_repeat = true;
-    }
-    else
-    {
-      // キーアップ時
-      if (is_swiped == false)
-      {
-        tap_code(keycode);
-      }
-
-      disable_click_layer();
-      is_swiped = false;
-      is_repeat = false;
-    }
-    return false;
-
-  default:
-    if (record->event.pressed)
-    {
-      // キーダウン時
-      disable_click_layer();
-    }
-  }
-
-  return true;
-}
-
-// ball_stateの設定
-report_mouse_t pointing_device_task_user(report_mouse_t mouse_report)
-{
-  int16_t current_x = mouse_report.x;
-  int16_t current_y = mouse_report.y;
-
-  if (current_x != 0 || current_y != 0)
-  {
-
-    switch (state)
-    {
-    case CLICKABLE:
-      click_timer = timer_read();
-      break;
-
-    case CLICKING:
-      break;
-
-    case WAITING:
-      mouse_movement += my_abs(current_x) + my_abs(current_y);
-
-      if (mouse_movement >= to_clickable_movement)
-      {
-        mouse_movement = 0;
-        enable_click_layer();
-      }
-      break;
-
-    case SWIPE:
-      click_timer = timer_read();
-
-      if (my_abs(current_x) >= SWIPE_THRESHOLD || my_abs(current_y) >= SWIPE_THRESHOLD)
-      {
-        rgblight_sethsv(HSV_BLUE);
-        process_swipe_gesture(current_x, current_y);
-        is_swiped = true;
-
-        if (is_repeat == false)
-        {
-          state = SWIPING;
-        }
-      }
-      break;
-      ;
-
-    case SWIPING:
-      break;
-
-    default:
-      click_timer = timer_read();
-      state = WAITING;
-      mouse_movement = 0;
-    }
-  }
-  else
-  {
-    switch (state)
-    {
-    case CLICKING:
-      break;
-
-    case CLICKABLE:
-      if (timer_elapsed(click_timer) > to_reset_time)
-      {
-        disable_click_layer();
-      }
-      break;
-
-    case WAITING:
-      if (timer_elapsed(click_timer) > 50)
-      {
-        mouse_movement = 0;
-        state = NONE;
-      }
-      break;
-
-    case SWIPE:
-      rgblight_sethsv(HSV_RED);
-      break;
-
-    case SWIPING:
-      if (timer_elapsed(click_timer) > 300)
-      {
-        state = SWIPE;
-      }
-      break;
-
-    default:
-      mouse_movement = 0;
-      state = NONE;
-    }
-  }
-  mouse_report.x = current_x;
-  mouse_report.y = current_y;
-
-  return mouse_report;
-}
+#include "features/my_functions.h"
+#include "features/swipe_gesture.h"
+#include "features/auto_click_layer.h"
+#include "features/my_macro_keys.h"
 
 // clang-format off
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
